+++ internal/model/dto/response/user_response.go
@@ -3,7 +3,7 @@ package response
 import "time"
 
 type UserResponse struct {
-	ID        int64                `xorm:"'id'" json:"id"`
+	ID        uint                 `xorm:"'id'" json:"id"`
 	Username  string               `xorm:"'username'" json:"username"`
 	Nickname  string               `xorm:"'nickname'" json:"nickname"`
 	Email     string               `xorm:"'email'" json:"email"`
@@ -14,18 +14,18 @@ type UserResponse struct {
 }
 
 type UserResponseWithTeamId struct {
-	ID        int64     `xorm:"'id'" json:"id"`
+	ID        uint      `xorm:"'id'" json:"id"`
 	Username  string    `xorm:"'username'" json:"username"`
 	Nickname  string    `xorm:"'nickname'" json:"nickname"`
 	Email     string    `xorm:"'email'" json:"email"`
 	Role      int64     `xorm:"'role'" json:"role"`
 	CreatedAt time.Time `xorm:"'created_at'" json:"created_at"`
 	UpdatedAt time.Time `xorm:"'updated_at'" json:"updated_at"`
-	TeamId    int64     `xorm:"'team_id'" json:"team_id"`
+	TeamId    uint      `xorm:"'team_id'" json:"team_id"`
 }
 
 type UserSimpleResponse struct {
-	ID       int64  `xorm:"'id'" json:"id"`
+	ID       uint   `xorm:"'id'" json:"id"`
 	Username string `xorm:"'username'" json:"username"`
 	Nickname string `xorm:"'nickname'" json:"nickname"`
 }
